Add files via upload
adding an another project to github cam eacross fibancci and golden numbers on youtube today, this morining and was facinated by how it works and wanted to learn about it so i practiced a simple code using for loop to make a fibonacci series for numbers. yet to explore the real worl applications of this mathematical wonder. enjoy the code i have description along the way